pyproject.toml
[project]
name = "podtool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "google-cloud-storage>=3.1.0",
    "openai>=1.66.3",
]
[project.scripts]
podtool = "podtool.main:cli"

[tool.setuptools]
packages = ["podtool"]

